refactor: started move of dump cmd
